+++ packages/elements/src/components/ui/modal/ModalContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
-import { RejectCommand, ResolveCommand } from "./UseModal";
+import { RejectCommand, ResolveCommand } from "./ModalCommands";
 
 export interface ModalContextValue<TResolveValue> {
   resolve: ResolveCommand<TResolveValue>;
